split discounts to sales, promotions, vouchers
