+++ saleor/tests/e2e/checkout/discounts/promotions/test_checkout_products_on_percentage_promotion.py
@@ -1,10 +1,10 @@
 import pytest
 
-from ...product.utils.preparing_product import prepare_product
-from ...sales.utils import create_sale, create_sale_channel_listing, sale_catalogues_add
-from ...shop.utils.preparing_shop import prepare_shop
-from ...utils import assign_permissions
-from ..utils import (
+from ....product.utils.preparing_product import prepare_product
+from ....promotions.utils import create_promotion, create_promotion_rule
+from ....shop.utils.preparing_shop import prepare_shop
+from ....utils import assign_permissions
+from ...utils import (
     checkout_complete,
     checkout_create,
     checkout_delivery_method_update,
@@ -12,44 +12,9 @@ from ..utils import (
 )
 
 
-def prepare_sale_for_product(
-    e2e_staff_api_client,
-    channel_id,
-    product_id,
-    sale_discount_type,
-    sale_discount_value,
-):
-    sale_name = "Sale"
-    sale = create_sale(
-        e2e_staff_api_client,
-        sale_name,
-        sale_discount_type,
-    )
-    sale_id = sale["id"]
-    sale_listing_input = [
-        {
-            "channelId": channel_id,
-            "discountValue": sale_discount_value,
-        }
-    ]
-    create_sale_channel_listing(
-        e2e_staff_api_client,
-        sale_id,
-        add_channels=sale_listing_input,
-    )
-    catalogue_input = {"products": [product_id]}
-    sale_catalogues_add(
-        e2e_staff_api_client,
-        sale_id,
-        catalogue_input,
-    )
-
-    return sale_id, sale_discount_value
-
-
 @pytest.mark.e2e
-def test_checkout_products_on_fixed_sale_core_1002(
-    e2e_not_logged_api_client,
+def test_checkout_products_on_percentage_promotion_core_2104(
+    e2e_logged_api_client,
     e2e_staff_api_client,
     permission_manage_products,
     permission_manage_channels,
@@ -58,6 +23,7 @@ def test_checkout_products_on_fixed_sale_core_1002(
     permission_manage_discounts,
 ):
     # Before
+
     permissions = [
         permission_manage_products,
         permission_manage_channels,
@@ -79,26 +45,38 @@ def test_checkout_products_on_fixed_sale_core_1002(
         product_variant_id,
         product_variant_price,
     ) = prepare_product(
-        e2e_staff_api_client,
-        warehouse_id,
-        channel_id,
-        variant_price="13.33",
+        e2e_staff_api_client, warehouse_id, channel_id, variant_price=19.89
     )
 
-    sale_id, sale_discount_value = prepare_sale_for_product(
-        e2e_staff_api_client,
-        channel_id,
-        product_id,
-        sale_discount_type="FIXED",
-        sale_discount_value=3,
-    )
+    promotion_name = "Promotion PERCENTAGE"
+    discount_value = 3
+    discount_type = "PERCENTAGE"
+    promotion_rule_name = "rule for product"
 
-    # Step 1 - checkoutCreate for product on sale
+    promotion_data = create_promotion(e2e_staff_api_client, promotion_name)
+    promotion_id = promotion_data["id"]
+
+    catalogue_predicate = {"productPredicate": {"ids": [product_id]}}
+
+    promotion_rule = create_promotion_rule(
+        e2e_staff_api_client,
+        promotion_id,
+        catalogue_predicate,
+        discount_type,
+        discount_value,
+        promotion_rule_name,
+        channel_id,
+    )
+    product_predicate = promotion_rule["cataloguePredicate"]["productPredicate"]["ids"]
+    assert promotion_rule["channels"][0]["id"] == channel_id
+    assert product_predicate[0] == product_id
+
+    # Step 1 - checkoutCreate for product on promotion
     lines = [
         {"variantId": product_variant_id, "quantity": 2},
     ]
     checkout_data = checkout_create(
-        e2e_not_logged_api_client,
+        e2e_logged_api_client,
         lines,
         channel_slug,
         email="testEmail@example.com",
@@ -108,7 +86,8 @@ def test_checkout_products_on_fixed_sale_core_1002(
     checkout_id = checkout_data["id"]
     checkout_lines = checkout_data["lines"][0]
     shipping_method_id = checkout_data["shippingMethods"][0]["id"]
-    unit_price = float(product_variant_price) - sale_discount_value
+    line_discount = round(float(product_variant_price) * discount_value / 100, 2)
+    unit_price = float(product_variant_price) - line_discount
 
     assert checkout_data["isShippingRequired"] is True
     assert checkout_lines["unitPrice"]["gross"]["amount"] == unit_price
@@ -118,35 +97,28 @@ def test_checkout_products_on_fixed_sale_core_1002(
 
     # Step 2 - Set DeliveryMethod for checkout.
     checkout_data = checkout_delivery_method_update(
-        e2e_not_logged_api_client,
+        e2e_logged_api_client,
         checkout_id,
         shipping_method_id,
     )
     assert checkout_data["deliveryMethod"]["id"] == shipping_method_id
     total_gross_amount = checkout_data["totalPrice"]["gross"]["amount"]
-    subtotal_gross_amount = checkout_data["subtotalPrice"]["gross"]["amount"]
 
     # Step 3 - Create payment for checkout.
     checkout_dummy_payment_create(
-        e2e_not_logged_api_client,
-        checkout_id,
-        total_gross_amount,
+        e2e_logged_api_client, checkout_id, total_gross_amount
     )
 
     # Step 5 - Complete checkout.
-    order_data = checkout_complete(
-        e2e_not_logged_api_client,
-        checkout_id,
-    )
-    assert order_data["status"] == "UNFULFILLED"
-    assert order_data["total"]["gross"]["amount"] == total_gross_amount
-    assert order_data["subtotal"]["gross"]["amount"] == subtotal_gross_amount
+    order_data = checkout_complete(e2e_logged_api_client, checkout_id)
 
     order_line = order_data["lines"][0]
+    assert order_data["status"] == "UNFULFILLED"
+    assert order_data["total"]["gross"]["amount"] == total_gross_amount
     assert order_line["undiscountedUnitPrice"]["gross"]["amount"] == float(
         product_variant_price
     )
     assert order_line["unitDiscountType"] == "FIXED"
     assert order_line["unitPrice"]["gross"]["amount"] == unit_price
-    assert order_line["unitDiscount"]["amount"] == float(sale_discount_value)
-    assert order_line["unitDiscountReason"] == f"Sale: {sale_id}"
+    assert order_line["unitDiscount"]["amount"] == line_discount
+    assert order_line["unitDiscountReason"] == f"Promotion: {promotion_id}"
