+++ saleor/tests/e2e/checkout/discounts/vouchers/test_checkout_unable_to_update_single_use_settings_after_usage.py
@@ -1,43 +1,37 @@
-import datetime
-
 import pytest
-from django.utils import timezone
 
-from ...product.utils.preparing_product import prepare_product
-from ...shop.utils.preparing_shop import prepare_shop
-from ...utils import assign_permissions
-from ...vouchers.utils import (
+from ....product.utils.preparing_product import prepare_product
+from ....shop.utils.preparing_shop import prepare_shop
+from ....utils import assign_permissions
+from ....vouchers.utils import (
     create_voucher,
     create_voucher_channel_listing,
-    get_voucher,
+    raw_update_voucher,
 )
-from ..utils import (
+from ...utils import (
+    checkout_add_promo_code,
     checkout_complete,
     checkout_create,
     checkout_delivery_method_update,
     checkout_dummy_payment_create,
-    raw_checkout_add_promo_code,
 )
 
 
 def prepare_voucher(
     e2e_staff_api_client,
     channel_id,
-    voucher_code,
+    voucher_code_list,
     voucher_discount_type,
     voucher_discount_value,
     voucher_type,
-    start_date,
 ):
-    input = {
-        "code": voucher_code,
+    input_data = {
+        "addCodes": voucher_code_list,
         "discountValueType": voucher_discount_type,
         "type": voucher_type,
-        "usageLimit": 1,
         "singleUse": True,
-        "startDate": start_date,
     }
-    voucher_data = create_voucher(e2e_staff_api_client, input)
+    voucher_data = create_voucher(e2e_staff_api_client, input_data)
 
     voucher_id = voucher_data["id"]
     channel_listing = [
@@ -52,12 +46,12 @@ def prepare_voucher(
         channel_listing,
     )
 
-    return voucher_code, voucher_id
+    return voucher_code_list, voucher_id
 
 
 @pytest.mark.e2e
-def test_checkout_unable_to_use_not_started_voucher_CORE_0920(
-    e2e_not_logged_api_client,
+def test_checkout_unable_to_update_single_use_settings_after_usage_CORE_0926(
+    e2e_logged_api_client,
     e2e_staff_api_client,
     permission_manage_products,
     permission_manage_channels,
@@ -94,29 +88,23 @@ def test_checkout_unable_to_use_not_started_voucher_CORE_0920(
         channel_id,
         variant_price=19.99,
     )
-
-    now = timezone.now()
-    date_in_the_future = now + datetime.timedelta(days=2)
-
-    voucher_code, voucher_id = prepare_voucher(
+    voucher_code_list, voucher_id = prepare_voucher(
         e2e_staff_api_client,
         channel_id,
-        voucher_code="percentage_voucher",
+        voucher_code_list=["single_use_1", "single_use_2"],
         voucher_discount_type="PERCENTAGE",
         voucher_discount_value=15,
         voucher_type="ENTIRE_ORDER",
-        start_date=date_in_the_future,
     )
-
-    # Step 1 - Create checkout for not logged in user
+    # Step 1 - Create checkout
     lines = [
         {
             "variantId": product_variant_id,
-            "quantity": 1,
+            "quantity": 2,
         },
     ]
     checkout = checkout_create(
-        e2e_not_logged_api_client,
+        e2e_logged_api_client,
         lines,
         channel_slug,
         email="testEmail@example.com",
@@ -127,46 +115,60 @@ def test_checkout_unable_to_use_not_started_voucher_CORE_0920(
     checkout_lines = checkout["lines"][0]
     shipping_method_id = checkout["shippingMethods"][0]["id"]
     unit_price = float(product_variant_price)
-    total_gross_amount = checkout["totalPrice"]["gross"]["amount"]
     assert checkout["isShippingRequired"] is True
     assert checkout_lines["unitPrice"]["gross"]["amount"] == unit_price
-    assert checkout_lines["undiscountedUnitPrice"]["amount"] == unit_price
+    assert checkout_lines["undiscountedUnitPrice"]["amount"] == float(
+        product_variant_price
+    )
 
     # Step 2 - Set DeliveryMethod for checkout.
     checkout_data = checkout_delivery_method_update(
-        e2e_not_logged_api_client,
+        e2e_logged_api_client,
         checkout_id,
         shipping_method_id,
     )
     assert checkout_data["deliveryMethod"]["id"] == shipping_method_id
     total_gross_amount = checkout_data["totalPrice"]["gross"]["amount"]
 
-    # Step 3 - Add voucher code to the checkout
-    data = raw_checkout_add_promo_code(
-        e2e_not_logged_api_client,
+    # Step 3 - Add voucher code to checkout
+    data = checkout_add_promo_code(
+        e2e_logged_api_client,
         checkout_id,
-        voucher_code,
+        voucher_code_list[0],
     )
-    errors = data["errors"][0]
-    assert errors["code"] == "INVALID"
-    assert errors["field"] == "promoCode"
-    assert errors["message"] == "Promo code is invalid"
-
-    voucher_data = get_voucher(e2e_staff_api_client, voucher_id)
-    assert voucher_data["voucher"]["id"] == voucher_id
-    assert voucher_data["voucher"]["codes"]["edges"][0]["node"]["used"] == 0
-
+    discounted_total_gross = data["totalPrice"]["gross"]["amount"]
+    discounted_unit_price = data["lines"][0]["unitPrice"]["gross"]["amount"]
+    voucher_discount = 2 * (round(float(product_variant_price) * 15 / 100, 2))
+    unit_price = float(product_variant_price)
+    assert data["discount"]["amount"] == voucher_discount
+    assert discounted_total_gross == total_gross_amount - voucher_discount
+    assert discounted_unit_price == unit_price - (
+        round(float(product_variant_price) * 15 / 100, 2)
+    )
     # Step 4 - Create payment for checkout.
     checkout_dummy_payment_create(
-        e2e_not_logged_api_client,
+        e2e_logged_api_client,
         checkout_id,
-        total_gross_amount,
+        discounted_total_gross,
     )
 
     # Step 5 - Complete checkout.
-    order_data = checkout_complete(e2e_not_logged_api_client, checkout_id)
+    order_data = checkout_complete(e2e_logged_api_client, checkout_id)
+
     order_line = order_data["lines"][0]
     assert order_data["status"] == "UNFULFILLED"
-    assert order_data["total"]["gross"]["amount"] == total_gross_amount
-    assert order_line["undiscountedUnitPrice"]["gross"]["amount"] == unit_price
-    assert order_data["discounts"] == []
+    assert order_data["total"]["gross"]["amount"] == discounted_total_gross
+    assert order_line["undiscountedUnitPrice"]["gross"]["amount"] == float(
+        product_variant_price
+    )
+    assert order_line["unitPrice"]["gross"]["amount"] == discounted_unit_price
+
+    # Step 6 - Update the voucher's single use settings
+    data = raw_update_voucher(e2e_staff_api_client, voucher_id, {"singleUse": False})
+    errors = data["errors"][0]
+    assert errors["code"] == "VOUCHER_ALREADY_USED"
+    assert errors["field"] == "singleUse"
+    assert (
+        errors["message"]
+        == "Cannot change single use setting when any voucher code has already been used."
+    )
