+++ saleor/tests/e2e/checkout/discounts/vouchers/test_checkout_unable_to_use_expired_voucher.py
@@ -1,30 +1,42 @@
-import pytest
+import datetime
 
-from ...product.utils.preparing_product import prepare_product
-from ...shop.utils.preparing_shop import prepare_shop
-from ...utils import assign_permissions
-from ...vouchers.utils import create_voucher, create_voucher_channel_listing
-from ..utils import (
-    checkout_add_promo_code,
+import pytest
+from django.utils import timezone
+
+from ....product.utils.preparing_product import prepare_product
+from ....shop.utils.preparing_shop import prepare_shop
+from ....utils import assign_permissions
+from ....vouchers.utils import (
+    create_voucher,
+    create_voucher_channel_listing,
+    get_voucher,
+)
+from ...utils import (
     checkout_complete,
     checkout_create,
     checkout_delivery_method_update,
     checkout_dummy_payment_create,
+    raw_checkout_add_promo_code,
 )
 
 
-def prepare_percentage_voucher(
+def prepare_voucher(
     e2e_staff_api_client,
     channel_id,
     voucher_code,
     voucher_discount_type,
     voucher_discount_value,
     voucher_type,
+    end_date,
 ):
     input = {
         "code": voucher_code,
         "discountValueType": voucher_discount_type,
         "type": voucher_type,
+        "usageLimit": 1,
+        "singleUse": True,
+        "startDate": "2023-10-01T18:26:00+02:00",
+        "endDate": end_date,
     }
     voucher_data = create_voucher(e2e_staff_api_client, input)
 
@@ -41,11 +53,11 @@ def prepare_percentage_voucher(
         channel_listing,
     )
 
-    return voucher_code
+    return voucher_code, voucher_id
 
 
 @pytest.mark.e2e
-def test_checkout_use_percentage_voucher_core_0902(
+def test_checkout_unable_to_use_expired_voucher_CORE_0921(
     e2e_logged_api_client,
     e2e_staff_api_client,
     permission_manage_products,
@@ -53,6 +65,7 @@ def test_checkout_use_percentage_voucher_core_0902(
     permission_manage_shipping,
     permission_manage_product_types_and_attributes,
     permission_manage_discounts,
+    permission_manage_checkouts,
 ):
     # Before
     permissions = [
@@ -61,6 +74,7 @@ def test_checkout_use_percentage_voucher_core_0902(
         permission_manage_shipping,
         permission_manage_product_types_and_attributes,
         permission_manage_discounts,
+        permission_manage_checkouts,
     ]
     assign_permissions(e2e_staff_api_client, permissions)
 
@@ -72,7 +86,7 @@ def test_checkout_use_percentage_voucher_core_0902(
     ) = prepare_shop(e2e_staff_api_client)
 
     (
-        _product_id,
+        product_id,
         product_variant_id,
         product_variant_price,
     ) = prepare_product(
@@ -82,20 +96,24 @@ def test_checkout_use_percentage_voucher_core_0902(
         variant_price=19.99,
     )
 
-    voucher_code = prepare_percentage_voucher(
+    now = timezone.now()
+    date_in_the_past = now - datetime.timedelta(days=2)
+
+    voucher_code, voucher_id = prepare_voucher(
         e2e_staff_api_client,
         channel_id,
-        voucher_code="PERCENTAGE_VOUCHER",
-        voucher_discount_type="PERCENTAGE",
-        voucher_discount_value=10,
+        voucher_code="fixed_voucher",
+        voucher_discount_type="FIXED",
+        voucher_discount_value=15,
         voucher_type="ENTIRE_ORDER",
+        end_date=date_in_the_past,
     )
 
-    # Step 1 - Create checkout
+    # Step 1 - Create checkout for not logged in user
     lines = [
         {
             "variantId": product_variant_id,
-            "quantity": 2,
+            "quantity": 1,
         },
     ]
     checkout = checkout_create(
@@ -110,11 +128,10 @@ def test_checkout_use_percentage_voucher_core_0902(
     checkout_lines = checkout["lines"][0]
     shipping_method_id = checkout["shippingMethods"][0]["id"]
     unit_price = float(product_variant_price)
+    total_gross_amount = checkout["totalPrice"]["gross"]["amount"]
     assert checkout["isShippingRequired"] is True
     assert checkout_lines["unitPrice"]["gross"]["amount"] == unit_price
-    assert checkout_lines["undiscountedUnitPrice"]["amount"] == float(
-        product_variant_price
-    )
+    assert checkout_lines["undiscountedUnitPrice"]["amount"] == unit_price
 
     # Step 2 - Set DeliveryMethod for checkout.
     checkout_data = checkout_delivery_method_update(
@@ -125,36 +142,32 @@ def test_checkout_use_percentage_voucher_core_0902(
     assert checkout_data["deliveryMethod"]["id"] == shipping_method_id
     total_gross_amount = checkout_data["totalPrice"]["gross"]["amount"]
 
-    # Step 3 - Add voucher code to checkout
-    data = checkout_add_promo_code(
+    # Step 3 - Add voucher code to the checkout
+    data = raw_checkout_add_promo_code(
         e2e_logged_api_client,
         checkout_id,
         voucher_code,
     )
-    discounted_total_gross = data["totalPrice"]["gross"]["amount"]
-    discounted_unit_price = data["lines"][0]["unitPrice"]["gross"]["amount"]
-    voucher_discount = 2 * (round(float(product_variant_price) * 10 / 100, 2))
+    errors = data["errors"][0]
+    assert errors["code"] == "INVALID"
+    assert errors["field"] == "promoCode"
+    assert errors["message"] == "Promo code is invalid"
 
-    assert data["discount"]["amount"] == voucher_discount
-    assert discounted_total_gross == total_gross_amount - voucher_discount
-    assert discounted_unit_price == unit_price - (
-        round(float(product_variant_price) * 10 / 100, 2)
-    )
+    voucher_data = get_voucher(e2e_staff_api_client, voucher_id)
+    assert voucher_data["voucher"]["id"] == voucher_id
+    assert voucher_data["voucher"]["codes"]["edges"][0]["node"]["used"] == 0
 
     # Step 4 - Create payment for checkout.
     checkout_dummy_payment_create(
         e2e_logged_api_client,
         checkout_id,
-        discounted_total_gross,
+        total_gross_amount,
     )
 
     # Step 5 - Complete checkout.
     order_data = checkout_complete(e2e_logged_api_client, checkout_id)
-
     order_line = order_data["lines"][0]
     assert order_data["status"] == "UNFULFILLED"
-    assert order_data["total"]["gross"]["amount"] == discounted_total_gross
-    assert order_line["undiscountedUnitPrice"]["gross"]["amount"] == float(
-        product_variant_price
-    )
-    assert order_line["unitPrice"]["gross"]["amount"] == discounted_unit_price
+    assert order_data["total"]["gross"]["amount"] == total_gross_amount
+    assert order_line["undiscountedUnitPrice"]["gross"]["amount"] == unit_price
+    assert order_data["discounts"] == []
