+++ saleor/tests/e2e/checkout/discounts/vouchers/test_checkout_user_cannot_use_voucher_for_staff_only.py
@@ -1,16 +1,10 @@
 import pytest
 
-from ...product.utils.preparing_product import prepare_product
-from ...shop.utils.preparing_shop import prepare_shop
-from ...utils import assign_permissions
-from ...vouchers.utils import create_voucher, create_voucher_channel_listing
-from ..utils import (
-    checkout_add_promo_code,
-    checkout_complete,
-    checkout_create,
-    checkout_delivery_method_update,
-    checkout_dummy_payment_create,
-)
+from ....product.utils.preparing_product import prepare_product
+from ....shop.utils.preparing_shop import prepare_shop
+from ....utils import assign_permissions
+from ....vouchers.utils import create_voucher, create_voucher_channel_listing
+from ...utils import checkout_create, raw_checkout_add_promo_code
 
 
 def prepare_voucher_for_staff_only(
@@ -45,9 +39,9 @@ def prepare_voucher_for_staff_only(
 
 
 @pytest.mark.e2e
-def test_staff_can_use_voucher_for_staff_only_in_checkout_core_0904(
+def test_user_cannot_use_voucher_for_staff_only_in_checkout_core_0905(
     e2e_staff_api_client,
-    e2e_no_permission_staff_api_client,
+    e2e_not_logged_api_client,
     permission_manage_products,
     permission_manage_channels,
     permission_manage_shipping,
@@ -69,7 +63,7 @@ def test_staff_can_use_voucher_for_staff_only_in_checkout_core_0904(
         warehouse_id,
         channel_id,
         channel_slug,
-        shipping_method_id,
+        _shipping_method_id,
     ) = prepare_shop(e2e_staff_api_client)
 
     (
@@ -80,14 +74,14 @@ def test_staff_can_use_voucher_for_staff_only_in_checkout_core_0904(
         e2e_staff_api_client,
         warehouse_id,
         channel_id,
-        variant_price="9.99",
+        variant_price="36.66",
     )
 
     voucher_discount_value, voucher_code = prepare_voucher_for_staff_only(
         e2e_staff_api_client,
         channel_id,
         "VOUCHER001",
-        "FIXED",
+        "PERCENTAGE",
         1,
     )
 
@@ -99,7 +93,7 @@ def test_staff_can_use_voucher_for_staff_only_in_checkout_core_0904(
         },
     ]
     checkout_data = checkout_create(
-        e2e_no_permission_staff_api_client,
+        e2e_not_logged_api_client,
         lines,
         channel_slug,
         email="testEmail@example.com",
@@ -107,7 +101,6 @@ def test_staff_can_use_voucher_for_staff_only_in_checkout_core_0904(
         set_default_shipping_address=True,
     )
     checkout_id = checkout_data["id"]
-    shipping_method_id = checkout_data["shippingMethods"][0]["id"]
     checkout_lines = checkout_data["lines"][0]
     assert checkout_lines["unitPrice"]["gross"]["amount"] == float(
         product_variant_price
@@ -115,43 +108,12 @@ def test_staff_can_use_voucher_for_staff_only_in_checkout_core_0904(
     assert checkout_data["isShippingRequired"] is True
 
     # Step 2 Add voucher code to checkout
-    checkout_data = checkout_add_promo_code(
-        e2e_no_permission_staff_api_client,
+    checkout_data = raw_checkout_add_promo_code(
+        e2e_not_logged_api_client,
         checkout_id,
         voucher_code,
     )
-    unit_price_with_voucher = float(product_variant_price) - voucher_discount_value
-    assert (
-        checkout_data["lines"][0]["unitPrice"]["gross"]["amount"]
-        == unit_price_with_voucher
-    )
-
-    # Step 4 - Set DeliveryMethod for checkout.
-    checkout_data = checkout_delivery_method_update(
-        e2e_no_permission_staff_api_client,
-        checkout_id,
-        shipping_method_id,
-    )
-    assert checkout_data["deliveryMethod"]["id"] == shipping_method_id
-    shipping_price = checkout_data["deliveryMethod"]["price"]["amount"]
-    total_gross_amount = round(unit_price_with_voucher + shipping_price, 2)
-    assert checkout_data["totalPrice"]["gross"]["amount"] == total_gross_amount
-
-    # Step 5 - Create payment for checkout.
-    checkout_dummy_payment_create(
-        e2e_no_permission_staff_api_client,
-        checkout_id,
-        total_gross_amount,
-    )
-
-    # Step 6 - Complete checkout.
-    order_data = checkout_complete(
-        e2e_no_permission_staff_api_client,
-        checkout_id,
-    )
-    assert order_data["status"] == "UNFULFILLED"
-    assert order_data["discounts"][0]["type"] == "VOUCHER"
-    assert order_data["voucher"]["code"] == voucher_code
-    assert order_data["discounts"][0]["value"] == voucher_discount_value
-    order_total_gross_amount = order_data["total"]["gross"]["amount"]
-    assert order_total_gross_amount == total_gross_amount
+    errors = checkout_data["errors"][0]
+    assert errors["code"] == "VOUCHER_NOT_APPLICABLE"
+    assert errors["field"] == "promoCode"
+    assert errors["message"] == "Voucher is not applicable to this checkout."
