+++ saleor/tests/e2e/checkout/discounts/vouchers/test_checkout_voucher_is_still_active_when_checkout_fails.py
@@ -1,22 +1,20 @@
-import datetime
-
 import pytest
-from django.utils import timezone
 
-from ...product.utils.preparing_product import prepare_product
-from ...shop.utils.preparing_shop import prepare_shop
-from ...utils import assign_permissions
-from ...vouchers.utils import (
+from ....product.utils import product_variant_stock_update
+from ....product.utils.preparing_product import prepare_product
+from ....shop.utils.preparing_shop import prepare_shop
+from ....utils import assign_permissions
+from ....vouchers.utils import (
     create_voucher,
     create_voucher_channel_listing,
     get_voucher,
 )
-from ..utils import (
-    checkout_complete,
+from ...utils import (
+    checkout_add_promo_code,
     checkout_create,
     checkout_delivery_method_update,
     checkout_dummy_payment_create,
-    raw_checkout_add_promo_code,
+    raw_checkout_complete,
 )
 
 
@@ -27,16 +25,11 @@ def prepare_voucher(
     voucher_discount_type,
     voucher_discount_value,
     voucher_type,
-    end_date,
 ):
     input = {
         "code": voucher_code,
         "discountValueType": voucher_discount_type,
         "type": voucher_type,
-        "usageLimit": 1,
-        "singleUse": True,
-        "startDate": "2023-10-01T18:26:00+02:00",
-        "endDate": end_date,
     }
     voucher_data = create_voucher(e2e_staff_api_client, input)
 
@@ -57,7 +50,7 @@ def prepare_voucher(
 
 
 @pytest.mark.e2e
-def test_checkout_unable_to_use_expired_voucher_CORE_0921(
+def test_checkout_voucher_is_still_active_when_checkout_fails_core_0918(
     e2e_logged_api_client,
     e2e_staff_api_client,
     permission_manage_products,
@@ -86,7 +79,7 @@ def test_checkout_unable_to_use_expired_voucher_CORE_0921(
     ) = prepare_shop(e2e_staff_api_client)
 
     (
-        product_id,
+        _product_id,
         product_variant_id,
         product_variant_price,
     ) = prepare_product(
@@ -96,24 +89,20 @@ def test_checkout_unable_to_use_expired_voucher_CORE_0921(
         variant_price=19.99,
     )
 
-    now = timezone.now()
-    date_in_the_past = now - datetime.timedelta(days=2)
-
     voucher_code, voucher_id = prepare_voucher(
         e2e_staff_api_client,
         channel_id,
-        voucher_code="fixed_voucher",
-        voucher_discount_type="FIXED",
-        voucher_discount_value=15,
+        voucher_code="PERCENTAGE_VOUCHER",
+        voucher_discount_type="PERCENTAGE",
+        voucher_discount_value=10,
         voucher_type="ENTIRE_ORDER",
-        end_date=date_in_the_past,
     )
 
-    # Step 1 - Create checkout for not logged in user
+    # Step 1 - Create checkout
     lines = [
         {
             "variantId": product_variant_id,
-            "quantity": 1,
+            "quantity": 2,
         },
     ]
     checkout = checkout_create(
@@ -129,9 +118,12 @@ def test_checkout_unable_to_use_expired_voucher_CORE_0921(
     shipping_method_id = checkout["shippingMethods"][0]["id"]
     unit_price = float(product_variant_price)
     total_gross_amount = checkout["totalPrice"]["gross"]["amount"]
+
     assert checkout["isShippingRequired"] is True
     assert checkout_lines["unitPrice"]["gross"]["amount"] == unit_price
-    assert checkout_lines["undiscountedUnitPrice"]["amount"] == unit_price
+    assert checkout_lines["undiscountedUnitPrice"]["amount"] == float(
+        product_variant_price
+    )
 
     # Step 2 - Set DeliveryMethod for checkout.
     checkout_data = checkout_delivery_method_update(
@@ -142,32 +134,44 @@ def test_checkout_unable_to_use_expired_voucher_CORE_0921(
     assert checkout_data["deliveryMethod"]["id"] == shipping_method_id
     total_gross_amount = checkout_data["totalPrice"]["gross"]["amount"]
 
-    # Step 3 - Add voucher code to the checkout
-    data = raw_checkout_add_promo_code(
+    # Step 3 - Add voucher code to checkout
+    data = checkout_add_promo_code(
         e2e_logged_api_client,
         checkout_id,
         voucher_code,
     )
-    errors = data["errors"][0]
-    assert errors["code"] == "INVALID"
-    assert errors["field"] == "promoCode"
-    assert errors["message"] == "Promo code is invalid"
+    discounted_total_gross = data["totalPrice"]["gross"]["amount"]
+    discounted_unit_price = data["lines"][0]["unitPrice"]["gross"]["amount"]
+    voucher_discount = 2 * (round(float(product_variant_price) * 10 / 100, 2))
 
-    voucher_data = get_voucher(e2e_staff_api_client, voucher_id)
-    assert voucher_data["voucher"]["id"] == voucher_id
-    assert voucher_data["voucher"]["codes"]["edges"][0]["node"]["used"] == 0
+    assert data["discount"]["amount"] == voucher_discount
+    assert discounted_total_gross == total_gross_amount - voucher_discount
+    assert discounted_unit_price == unit_price - (
+        round(float(product_variant_price) * 10 / 100, 2)
+    )
 
     # Step 4 - Create payment for checkout.
     checkout_dummy_payment_create(
         e2e_logged_api_client,
         checkout_id,
-        total_gross_amount,
+        discounted_total_gross,
+    )
+    # Step 5 - Change the product variant stock
+    data = product_variant_stock_update(
+        e2e_staff_api_client,
+        warehouse_id,
+        1,
+        product_variant_id,
     )
 
-    # Step 5 - Complete checkout.
-    order_data = checkout_complete(e2e_logged_api_client, checkout_id)
-    order_line = order_data["lines"][0]
-    assert order_data["status"] == "UNFULFILLED"
-    assert order_data["total"]["gross"]["amount"] == total_gross_amount
-    assert order_line["undiscountedUnitPrice"]["gross"]["amount"] == unit_price
-    assert order_data["discounts"] == []
+    # Step 6 - Complete checkout
+    order_data = raw_checkout_complete(e2e_logged_api_client, checkout_id)
+
+    assert order_data["errors"] != []
+    assert order_data["errors"][0]["code"] == "INSUFFICIENT_STOCK"
+
+    # Step 7 - Check the voucher code is still active and is assigned to checkout
+    voucher_data = get_voucher(e2e_staff_api_client, voucher_id)
+    assert voucher_data["voucher"]["id"] == voucher_id
+    assert voucher_data["voucher"]["codes"]["edges"][0]["node"]["used"] == 0
+    assert voucher_data["checkouts"]["edges"][0]["node"]["id"] == checkout_id
