+++ saleor/tests/e2e/checkout/discounts/vouchers/test_should_be_able_to_remove_voucher_code_from_checkout.py
@@ -1,15 +1,20 @@
 import pytest
 
-from ...product.utils.preparing_product import prepare_product
-from ...shop.utils.preparing_shop import prepare_shop
-from ...utils import assign_permissions
-from ...vouchers.utils import create_voucher, create_voucher_channel_listing
-from ..utils import (
+from ....product.utils.preparing_product import prepare_product
+from ....shop.utils.preparing_shop import prepare_shop
+from ....utils import assign_permissions
+from ....vouchers.utils import (
+    create_voucher,
+    create_voucher_channel_listing,
+    get_voucher,
+)
+from ...utils import (
     checkout_add_promo_code,
     checkout_complete,
     checkout_create,
     checkout_delivery_method_update,
     checkout_dummy_payment_create,
+    checkout_remove_promo_code,
     checkout_shipping_address_update,
 )
 
@@ -21,13 +26,11 @@ def prepare_fixed_voucher(
     voucher_discount_type,
     voucher_discount_value,
     voucher_type,
-    products,
 ):
     input = {
         "code": voucher_code,
         "discountValueType": voucher_discount_type,
         "type": voucher_type,
-        "products": products,
     }
     voucher_data = create_voucher(e2e_staff_api_client, input)
 
@@ -44,11 +47,11 @@ def prepare_fixed_voucher(
         channel_listing,
     )
 
-    return voucher_code
+    return voucher_code, voucher_id
 
 
 @pytest.mark.e2e
-def test_checkout_voucher_should_not_cause_negative_variant_price_CORE_0911(
+def test_should_be_able_to_remove_voucher_code_from_checkout_CORE_0917(
     e2e_not_logged_api_client,
     e2e_staff_api_client,
     permission_manage_products,
@@ -56,6 +59,7 @@ def test_checkout_voucher_should_not_cause_negative_variant_price_CORE_0911(
     permission_manage_shipping,
     permission_manage_product_types_and_attributes,
     permission_manage_discounts,
+    permission_manage_checkouts,
 ):
     # Before
     permissions = [
@@ -64,6 +68,7 @@ def test_checkout_voucher_should_not_cause_negative_variant_price_CORE_0911(
         permission_manage_shipping,
         permission_manage_product_types_and_attributes,
         permission_manage_discounts,
+        permission_manage_checkouts,
     ]
     assign_permissions(e2e_staff_api_client, permissions)
 
@@ -75,31 +80,30 @@ def test_checkout_voucher_should_not_cause_negative_variant_price_CORE_0911(
     ) = prepare_shop(e2e_staff_api_client)
 
     (
-        product_id,
+        _product_id,
         product_variant_id,
         product_variant_price,
     ) = prepare_product(
         e2e_staff_api_client,
         warehouse_id,
         channel_id,
-        variant_price=10,
+        variant_price=30,
     )
 
-    voucher_code = prepare_fixed_voucher(
+    voucher_code, voucher_id = prepare_fixed_voucher(
         e2e_staff_api_client,
         channel_id,
         voucher_code="FIXED_VOUCHER",
         voucher_discount_type="FIXED",
-        voucher_discount_value=20,
-        voucher_type="SPECIFIC_PRODUCT",
-        products=product_id,
+        voucher_discount_value=10,
+        voucher_type="ENTIRE_ORDER",
     )
 
     # Step 1 - Create checkout
     lines = [
         {
             "variantId": product_variant_id,
-            "quantity": 1,
+            "quantity": 2,
         },
     ]
     checkout = checkout_create(
@@ -108,13 +112,14 @@ def test_checkout_voucher_should_not_cause_negative_variant_price_CORE_0911(
         channel_slug,
         email="testEmail@example.com",
         set_default_billing_address=True,
+        set_default_shipping_address=True,
     )
     checkout_id = checkout["id"]
     checkout_lines = checkout["lines"][0]
     subtotal_amount = float(product_variant_price)
     assert checkout["isShippingRequired"] is True
     total_without_shipping = checkout["totalPrice"]["gross"]["amount"]
-    assert total_without_shipping == subtotal_amount
+    assert total_without_shipping == subtotal_amount * 2
     assert checkout_lines["unitPrice"]["gross"]["amount"] == subtotal_amount
     assert checkout_lines["undiscountedUnitPrice"]["amount"] == product_variant_price
 
@@ -142,27 +147,42 @@ def test_checkout_voucher_should_not_cause_negative_variant_price_CORE_0911(
         voucher_code,
     )
     discounted_total_gross = data["totalPrice"]["gross"]["amount"]
-    subtotal = data["subtotalPrice"]["gross"]["amount"]
-    assert subtotal == 0
-    assert discounted_total_gross == shipping_price
+    discount_value = data["discount"]["amount"]
+    assert discounted_total_gross == undiscounted_total_gross - discount_value
 
-    # Step 4 - Create payment for checkout.
+    # Step 4 - Remove the voucher code from the checkout
+    data = checkout_remove_promo_code(
+        e2e_not_logged_api_client,
+        checkout_id,
+        voucher_code,
+    )
+    assert data["checkout"]["voucherCode"] is None
+    undiscounted_total_price = data["checkout"]["totalPrice"]["gross"]["amount"]
+
+    # Step 5 - Create payment for checkout
     checkout_dummy_payment_create(
         e2e_not_logged_api_client,
         checkout_id,
-        discounted_total_gross,
+        undiscounted_total_price,
     )
 
-    # Step 5 - Complete checkout.
+    # Step 6 - Complete checkout
     order_data = checkout_complete(
         e2e_staff_api_client,
         checkout_id,
     )
     assert order_data["status"] == "UNFULFILLED"
-    assert order_data["discounts"][0]["type"] == "VOUCHER"
-    assert order_data["voucher"]["code"] == voucher_code
-    assert order_data["total"]["gross"]["amount"] == discounted_total_gross
+    assert order_data["discounts"] == []
+    assert order_data["voucher"] is None
+    assert order_data["total"]["gross"]["amount"] == undiscounted_total_gross
     assert order_data["deliveryMethod"]["id"] == shipping_method_id
     assert order_data["shippingPrice"]["gross"]["amount"] == shipping_price
     order_line = order_data["lines"][0]
-    assert order_line["unitPrice"]["gross"]["amount"] == 0
+    assert order_line["unitPrice"]["gross"]["amount"] == float(product_variant_price)
+    assert order_data["total"]["gross"]["amount"] != discounted_total_gross
+
+    # Step 7 - Check the voucher code has not been used
+    voucher_data = get_voucher(e2e_staff_api_client, voucher_id)
+    assert voucher_data["voucher"]["id"] == voucher_id
+    assert voucher_data["voucher"]["codes"]["edges"][0]["node"]["isActive"] is True
+    assert voucher_data["voucher"]["codes"]["edges"][0]["node"]["used"] == 0
